Improve inference for PhiNode.edges

+++ base/compiler/ssair/inlining.jl
@@ -569,7 +569,7 @@ function batch_inline!(todo::Vector{Any}, ir::IRCode, linetable::Vector{LineInfo
             elseif isa(stmt, GotoIfNot)
                 compact[idx] = GotoIfNot(stmt.cond, state.bb_rename[stmt.dest])
             elseif isa(stmt, PhiNode)
-                compact[idx] = PhiNode(Any[edge == length(state.bb_rename) ? length(state.new_cfg_blocks) : state.bb_rename[edge+1]-1 for edge in stmt.edges], stmt.values)
+                compact[idx] = PhiNode(Any[edge::Int == length(state.bb_rename) ? length(state.new_cfg_blocks) : state.bb_rename[edge::Int+1]-1 for edge in stmt.edges], stmt.values)
             end
         end
